WIP: current state of work; meeting with supervisor; cosmetics
