+++ matscipy/neighbours.py
@@ -19,10 +19,7 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 # ======================================================================
 
-import itertools
-
 import numpy as np
-from array import array
 
 from ase.data import atomic_numbers
 
@@ -30,6 +27,7 @@ import _matscipy
 from _matscipy import first_neighbours, triplet_list, first_triplets
 
 ###
+
 
 def mic(dr, cell, pbc=None):
     """
@@ -55,7 +53,7 @@ def mic(dr, cell, pbc=None):
     # periodic boundaries then and need to be unwrapped.
     rec = np.linalg.inv(cell)
     if pbc is not None:
-        rec *= np.array(pbc, dtype=int).reshape(3,1)
+        rec *= np.array(pbc, dtype=int).reshape(3, 1)
     dri = np.round(np.dot(dr, rec))
 
     # Unwrap
@@ -70,7 +68,7 @@ def neighbour_list(quantities, atoms=None, cutoff=None, positions=None,
     are included in the neighbor list but the complexity of neighbor list search
     for those can become n^2.
 
-    The neighbor list is sorted by first atom index 'i', but not by second 
+    The neighbor list is sorted by first atom index 'i', but not by second
     atom index 'j'.
 
     The neighbour list accepts either an ASE Atoms object or positions and cell
@@ -127,7 +125,7 @@ def neighbour_list(quantities, atoms=None, cutoff=None, positions=None,
 
     2. Coordination counting with different cutoffs for each pair of species
         i = neighbor_list('i', a,
-                           {('H', 'H'): 1.1, ('C', 'H'): 1.3, ('C', 'C'): 1.85})
+                          {('H', 'H'): 1.1, ('C', 'H'): 1.3, ('C', 'C'): 1.85})
         coord = np.bincount(i)
 
     3. Pair distribution function:
@@ -188,7 +186,7 @@ e_nc = (dr_nc.T/abs_dr_n).T
     if atoms is None:
         if positions is None:
             raise ValueError('You provided neither an ASE Atoms object nor '
-                             'a positions array.') 
+                             'a positions array.')
         if cell is None:
             # Shrink wrapped cell
             rmin = np.min(positions, axis=0)
@@ -255,7 +253,7 @@ def find_indices_of_reversed_pairs(i_n, j_n, abs_dr_n):
     determines the list of indices `reverse` into the neighbor list
     where each pair is reversed, i.e. `i_n[reverse[n]]=j_n[n]` and
     `j_n[reverse[n]]=i_n[n]` for each index `n` in the neighbor list
-    
+
     In the case of small periodic systems, one needs to be careful, because
     the same pair may appear more than one time, with different pair
     distances. Therefore, the pair distance must be taken into account.
@@ -279,10 +277,10 @@ def find_indices_of_reversed_pairs(i_n, j_n, abs_dr_n):
     """
     sorted_1 = np.lexsort(keys=(abs_dr_n, i_n, j_n))
     sorted_2 = np.lexsort(keys=(abs_dr_n, j_n, i_n))
-    #np.testing.assert_equal(j_n[sorted_1], i_n[sorted_2])
-    #np.testing.assert_equal(i_n[sorted_1], j_n[sorted_2])
-    #np.testing.assert_equal(abs_dr_n[sorted_1], abs_dr_n[sorted_2])
-    #print(np.c_[i_n[sorted_2], j_n[sorted_2], abs_dr_n[sorted_2], 
+    # np.testing.assert_equal(j_n[sorted_1], i_n[sorted_2])
+    # np.testing.assert_equal(i_n[sorted_1], j_n[sorted_2])
+    # np.testing.assert_equal(abs_dr_n[sorted_1], abs_dr_n[sorted_2])
+    # print(np.c_[i_n[sorted_2], j_n[sorted_2], abs_dr_n[sorted_2],
     #            i_n[sorted_1], j_n[sorted_1], abs_dr_n[sorted_1]])
     tmp2 = np.arange(i_n.size)[sorted_2]
     tmp1 = np.arange(i_n.size)[sorted_1]
@@ -292,101 +290,3 @@ def find_indices_of_reversed_pairs(i_n, j_n, abs_dr_n):
     reverse = np.empty(i_n.size, dtype=i_n.dtype)
     reverse[tmp1] = tmp2
     return reverse
-
-
-def triplet_list_py(first_neighbours):
-    """
-    Compute a triple list for an atomic configuration. The triple list is a
-    mask that can be applied to the corresponding neighbour list to mask
-    triplet properties.
-
-    The trplet list accepts an first_neighbour array (generated by
-    first_neighbours) as input.
-
-    Parameters
-    ----------
-    first_neighbours : array
-        adresses of the first time an atom occours in the neighour list
-
-    Returns
-    -------
-    ij_t, ik_t : array
-        lists of adresses that form triples in the pair lists
-
-    first_ij : array
-        adresses of the first triplet of a pair ij in the ij_t, ik_t lists
-
-    Examples
-    --------
-    Examples assume Atoms object *a* and numpy imported as *np*.
-    1. Calculate the angle between all triplets
-        #TODO
-
-    """
-    
-    """
-    # make sure to compute i_n since it's needed for the first_neighbours
-    i_n_pos = quantities.find('i')
-    if i_n_pos == -1:
-        i_n, neighbor_results = neighbour_list('i' + quantities, atoms, cutoff,
-                positions, cell, pbc, numbers, cell_origin)
-    else:
-        neighbor_results = neighbour_list(quantities, atoms, cutoff, positions,
-                cell, pbc, numbers, cell_origin)
-        i_n = neighbor_results[i_n_pos]
-
-    # compute the first_neighbours indicies and store it in first_i
-    first_i = first_neighbours(len(atoms), i_n)
-    """
-
-    """
-    # initialize ij_t and ik_t
-    ij_t = np.zeros(0, dtype=int)
-    ik_t = np.zeros(0, dtype=int)
-
-    # compute the triple list by processing first_i
-    diffs = np.diff(first_i)
-    for i in range(len(diffs)):
-        for r in range(1, diffs[i]):
-            ij_t = np.concatenate([ij_t,
-                                   np.ones(diffs[i]-r, dtype=int)*(first_i[i]+r-1)])
-            ik_t = np.concatenate([ik_t, np.arange(first_i[i],
-                                   first_i[i+1], dtype=int)[-(diffs[i]-r):]])
-    """
-
-    # newer approach
-    ij_t = np.zeros(0, dtype='int32')
-    ik_t = np.zeros(0, dtype='int32')
-    # because numpy.append is super slow (l. 410ff), array is used instead
-    # first_triplets = array('I', [0]) 
-    # first_triplets = np.zeros(1, dtype=int)
-
-    # compute the triple list by processing first_i
-    diffs = np.diff(first_neighbours)
-    for i in range(len(diffs)):
-        for r in range(1, diffs[i]+1):
-            ij_t = np.concatenate([ij_t,
-                    np.ones(diffs[i]-1, dtype='int32')*(first_neighbours[i]+r-1)])
-            ik_t = np.concatenate([ik_t, np.arange(first_neighbours[i],
-                                first_neighbours[i+1],
-                                dtype='int32')[np.arange(diffs[i])!=(r-1)].copy()])
-            # compute first_ij list
-            # first_triplets = np.append(first_triplets, [len(ij_t)])
-            # first_triplets.append(ij_t.size)
-            # first_triplets = np.concatenate([first_triplets, [len(ij_t)]])
-
-
-    return ij_t, ik_t
-
-def first_triplets_py(ij_t):
-    """
-    description
-    """
-    curr_neighbour_pair = 0
-    first_triplets  = np.zeros(1, dtype=int)
-    for i, curr in enumerate(ij_t):
-        if curr_neighbour_pair != curr:
-            curr_neighbour_pair = curr
-            first_triplets = np.concatenate([first_triplets, [i]])
-    first_triplets = np.concatenate([first_triplets, [ij_t.size]])
-    return first_triplets
