+++ setup.py
@@ -260,7 +260,7 @@ setup(name='matscipy',
                 'matscipy.calculators',
                 'matscipy.calculators.eam','matscipy.io','matscipy.tool',
                 'matscipy.calculators.pair_potential',
-                'matscipy.calculators.EDIPTersoffBrenner',
+                'matscipy.calculators.AbellTersoffBrenner',
                 'matscipy.calculators.mcfm',
                 'matscipy.calculators.mcfm.mcfm_parallel',
                 'matscipy.calculators.mcfm.neighbour_list_mcfm',
